add errors
